+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ##### Enhancements
 
-* Update commments/docs to indicate :skip will skip pch generation  
+* Update commments/docs to indicate prefix_header=false will skip pch generation  
   [Paul Beusterien](https://github.com/paulb777)
   [#412](https://github.com/CocoaPods/Core/pull/412)